requirements.txt
# --- file: requirements.txt ---
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.21.0
matplotlib>=3.4.0
seaborn>=0.11.0
scikit-learn>=1.0.0
Pillow>=8.3.0
efficientnet_pytorch>=0.7.0
timm>=0.6.0
pytorch_pretrained_vit>=0.0.7